Superbuild tl_common helper | TLOC_INCLUDE helper
Superbuild now writes a tl_include_common.cmake file in the root
directory to facilitate the inclusion of tl_base_ci's common header that
has all the basis functionality.

tl_common now has a TLOC_INCLUDE function that builds on include() by
searching for a file in a list of directories and including the first
one it finds.